moved LoggerThread to separate file
